Create Task redux + db

+++ to-do-app/src/components/SubjectsList/SubjectsList.js
@@ -1,19 +1,21 @@
 import React from 'react';
 import SubjectListItem from './SubjectListItem/SubjectListItem'
 const SubjectsList = (props) => {
-  console.log(props.arrList);
+  console.log("RECIEVED ITEMS IN SUBJECTLIST: ", props.arrList);
   let listItems = 'Loading';
-   listItems = props.arrList.map((item,index) => {
-     console.log('Index: ' + index)
-     return (
-        <SubjectListItem  
-          indexItem={index} 
-          name={item} 
-          deleteItem={props.deleteHandler} 
-          updateItem ={props.updateHandler}  
-          completeItem ={props.completeHandler}/>
-     )
-   });
+  let arr = props.arrList['tasks']
+  if(arr) {
+    listItems = arr.map((item) => {
+      return (
+         <SubjectListItem  
+           indexItem={item._id} 
+           name={item.name} 
+           deleteItem={props.deleteHandler} 
+           updateItem ={props.updateHandler}  
+           completeItem ={props.completeHandler}/>
+      )
+    });
+  }
   return ( 
     <div className="mt-8">{listItems}</div>
    );
